+++ app/src/main/java/com/example/ivan/champy_v2/utils/Constants.java
@@ -4,7 +4,7 @@ public class Constants {
 
     public static final String API_URL = "http://46.101.213.24:3007";
 
-    public static final long oneDay = 86400L;
+    public static final int oneDay = 86400;
 
     public static long unixTime = System.currentTimeMillis() / 1000L;
